refactor(execution): carve the result vocabulary into its own module (#2314)
Step 1 of the decomposition. `TaskExecutionErrorCode`, `TaskExecutionResult`
and `TerminalEnvelope` are the one part of task_execution_service with no
collaborators at all — pure dataclasses and an enum, no db, no HTTP, no
capacity manager — so they can be imported from anywhere without a cycle.
That is what lets the phase modules beside them name the same vocabulary
without importing the service they belong to.

Re-exported from task_execution_service: ~80 call sites and 83 test files
import these names from there, and a decomposition that renames the import
surface is not a pure refactor.
